Add PathPlanner settings, update swerve module
Back left swerve module had the same CAN ID as another device.

Added a missing settings file for PathPlanner, which would cause the robot to crash if this file was not present.

Added minimal SimGUI for simulator usage. YAGSL adds some NT providers to be used during runtime of the robot and simulator.