Adjust power/ drive curves
Smart current limit has been dropped from 50 amps to 45 to help with driver control.

Added a new drive power curve that halves the power of the swerve to allow for finer control. This is now default.

Added a button to activate the old drive power curve on a single button press. This by default is bound to the X button on the driver's Xbox controller.